# api-docs/swagger.yaml
swagger: "2.0"
info:
  title: API Bridge Service
  version: 1.0.0
  description: |
    API Bridge Service for Legacy and Modern System Integration
    
    이 서비스는 레거시 시스템과 모던 시스템 간의 API 브리지를 제공합니다.
    - API 라우팅 및 전환
    - 병렬 호출 및 응답 비교
    - 점진적 마이그레이션 지원
  termsOfService: http://swagger.io/terms/
  contact:
    name: API Support
    url: http://www.swagger.io/support
    email: support@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
host: localhost:10019
basePath: /
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: Authorization
...
paths:
...
  /api/v1/endpoints:
    get:
      tags:
        - Endpoints
      summary: List Endpoints
      description: 모든 API 엔드포인트 목록을 조회합니다
      operationId: listEndpoints
      produces:
        - application/json
      responses:
        "200":
          description: 엔드포인트 목록 조회 성공
          schema:
            type: object
            properties:
              endpoints:
                type: array
                items:
                  $ref: "#/definitions/EndpointResponse"
              count:
                type: integer
                example: 10
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    post:
      tags:
        - Endpoints
      summary: Create Endpoint
      description: 새로운 API 엔드포인트를 생성합니다
      operationId: createEndpoint
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: endpoint
          in: body
          description: 엔드포인트 정보
          required: true
          schema:
            $ref: "#/definitions/CreateEndpointRequest"
      responses:
        "201":
          description: 엔드포인트 생성 성공
          schema:
            $ref: "#/definitions/EndpointResponse"
        "400":
          description: 잘못된 요청
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/endpoints/{id}:
    get:
      tags:
        - Endpoints
      summary: Get Endpoint
      description: 특정 API 엔드포인트를 조회합니다
      operationId: getEndpoint
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 엔드포인트 ID
          required: true
          type: string
      responses:
        "200":
          description: 엔드포인트 조회 성공
          schema:
            $ref: "#/definitions/EndpointResponse"
        "404":
          description: 엔드포인트를 찾을 수 없음
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    put:
      tags:
        - Endpoints
      summary: Update Endpoint
      description: 기존 API 엔드포인트를 수정합니다
      operationId: updateEndpoint
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 엔드포인트 ID
          required: true
          type: string
        - name: endpoint
          in: body
          description: 수정할 엔드포인트 정보
          required: true
          schema:
            $ref: "#/definitions/UpdateEndpointRequest"
      responses:
        "200":
          description: 엔드포인트 수정 성공
          schema:
            $ref: "#/definitions/EndpointResponse"
        "400":
          description: 잘못된 요청
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: 엔드포인트를 찾을 수 없음
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      tags:
        - Endpoints
      summary: Delete Endpoint
      description: API 엔드포인트를 삭제합니다
      operationId: deleteEndpoint
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 엔드포인트 ID
          required: true
          type: string
      responses:
        "204":
          description: 엔드포인트 삭제 성공
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/routing-rules:
    get:
      tags:
        - Routing Rules
      summary: List Routing Rules
      description: 모든 라우팅 규칙 목록을 조회합니다
      operationId: listRoutingRules
      produces:
        - application/json
      responses:
        "200":
          description: 라우팅 규칙 목록 조회 성공
          schema:
            type: object
            properties:
              routing_rules:
                type: array
                items:
                  $ref: "#/definitions/RoutingRuleResponse"
              count:
                type: integer
                example: 5
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    post:
      tags:
        - Routing Rules
      summary: Create Routing Rule
      description: 새로운 라우팅 규칙을 생성합니다
      operationId: createRoutingRule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: routingRule
          in: body
          description: 라우팅 규칙 정보
          required: true
          schema:
            $ref: "#/definitions/CreateRoutingRuleRequest"
      responses:
        "201":
          description: 라우팅 규칙 생성 성공
          schema:
            $ref: "#/definitions/RoutingRuleResponse"
        "400":
          description: 잘못된 요청
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/routing-rules/{id}:
    get:
      tags:
        - Routing Rules
      summary: Get Routing Rule
      description: 특정 라우팅 규칙을 조회합니다
      operationId: getRoutingRule
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 라우팅 규칙 ID
          required: true
          type: string
      responses:
        "200":
          description: 라우팅 규칙 조회 성공
          schema:
            $ref: "#/definitions/RoutingRuleResponse"
        "404":
          description: 라우팅 규칙을 찾을 수 없음
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    put:
      tags:
        - Routing Rules
      summary: Update Routing Rule
      description: 기존 라우팅 규칙을 수정합니다
      operationId: updateRoutingRule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 라우팅 규칙 ID
          required: true
          type: string
        - name: routingRule
          in: body
          description: 수정할 라우팅 규칙 정보
          required: true
          schema:
            $ref: "#/definitions/UpdateRoutingRuleRequest"
      responses:
        "200":
          description: 라우팅 규칙 수정 성공
          schema:
            $ref: "#/definitions/RoutingRuleResponse"
        "400":
          description: 잘못된 요청
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: 라우팅 규칙을 찾을 수 없음
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    delete:
      tags:
        - Routing Rules
      summary: Delete Routing Rule
      description: 라우팅 규칙을 삭제합니다
      operationId: deleteRoutingRule
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 라우팅 규칙 ID
          required: true
          type: string
      responses:
        "204":
          description: 라우팅 규칙 삭제 성공
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/orchestration-rules:
    post:
      tags:
        - Orchestration Rules
      summary: Create Orchestration Rule
      description: 새로운 오케스트레이션 규칙을 생성합니다
      operationId: createOrchestrationRule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: orchestrationRule
          in: body
          description: 오케스트레이션 규칙 정보
          required: true
          schema:
            $ref: "#/definitions/CreateOrchestrationRuleRequest"
      responses:
        "201":
          description: 오케스트레이션 규칙 생성 성공
          schema:
            $ref: "#/definitions/OrchestrationRuleResponse"
        "400":
          description: 잘못된 요청
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/orchestration-rules/{id}:
    get:
      tags:
        - Orchestration Rules
      summary: Get Orchestration Rule
      description: 특정 오케스트레이션 규칙을 조회합니다
      operationId: getOrchestrationRule
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 오케스트레이션 규칙 ID
          required: true
          type: string
      responses:
        "200":
          description: 오케스트레이션 규칙 조회 성공
          schema:
            $ref: "#/definitions/OrchestrationRuleResponse"
        "404":
          description: 오케스트레이션 규칙을 찾을 수 없음
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
    
    put:
      tags:
        - Orchestration Rules
      summary: Update Orchestration Rule
      description: 기존 오케스트레이션 규칙을 수정합니다
      operationId: updateOrchestrationRule
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 오케스트레이션 규칙 ID
          required: true
          type: string
        - name: orchestrationRule
          in: body
          description: 수정할 오케스트레이션 규칙 정보
          required: true
          schema:
            $ref: "#/definitions/UpdateOrchestrationRuleRequest"
      responses:
        "200":
          description: 오케스트레이션 규칙 수정 성공
          schema:
            $ref: "#/definitions/OrchestrationRuleResponse"
        "400":
          description: 잘못된 요청
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: 오케스트레이션 규칙을 찾을 수 없음
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/orchestration-rules/{id}/evaluate-transition:
    get:
      tags:
        - Orchestration Rules
      summary: Evaluate Transition
      description: 오케스트레이션 규칙의 전환 가능 여부를 평가합니다
      operationId: evaluateTransition
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 오케스트레이션 규칙 ID
          required: true
          type: string
      responses:
        "200":
          description: 전환 평가 완료
          schema:
            type: object
            properties:
              can_transition:
                type: boolean
                example: true
                description: 전환 가능 여부
              current_mode:
                type: string
                example: PARALLEL
                description: 현재 모드
              rule_id:
                type: string
                example: orch-20240101120000-abc123
                description: 규칙 ID
        "404":
          description: 오케스트레이션 규칙을 찾을 수 없음
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/orchestration-rules/{id}/execute-transition:
    post:
      tags:
        - Orchestration Rules
      summary: Execute Transition
      description: 오케스트레이션 규칙의 API 모드를 전환합니다
      operationId: executeTransition
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: 오케스트레이션 규칙 ID
          required: true
          type: string
        - name: transition
          in: body
          description: 전환 정보
          required: true
          schema:
            type: object
            required:
              - new_mode
            properties:
              new_mode:
                type: string
                enum:
                  - LEGACY_ONLY
                  - MODERN_ONLY
                  - PARALLEL
                example: MODERN_ONLY
                description: 전환할 새로운 모드
      responses:
        "200":
          description: 전환 실행 성공
          schema:
            type: object
            properties:
              message:
                type: string
                example: transition executed successfully
              from_mode:
                type: string
                example: PARALLEL
              to_mode:
                type: string
                example: MODERN_ONLY
              rule_id:
                type: string
                example: orch-20240101120000-abc123
        "400":
          description: 잘못된 요청
          schema:
            $ref: "#/definitions/ErrorResponse"
        "404":
          description: 오케스트레이션 규칙을 찾을 수 없음
          schema:
            $ref: "#/definitions/ErrorResponse"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/v1/shutdown:
    post:
      tags:
        - System
      summary: Graceful Shutdown
      description: API Bridge 서비스를 안전하게 종료합니다
      operationId: gracefulShutdown
      consumes:
        - application/json
      produces:
        - application/json
      responses:
        "200":
          description: 종료 시작됨
          schema:
            type: object
            properties:
              message:
                type: string
                example: Graceful shutdown initiated
              timestamp:
                type: string
                format: date-time
                example: "2024-01-01T00:00:00Z"
        "500":
          description: 내부 서버 오류
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  CreateEndpointRequest:
    type: object
    required:
      - name
      - base_url
      - method
    properties:
      name:
        type: string
        description: 엔드포인트 이름
        example: Legacy User API
      description:
        type: string
        description: 엔드포인트 설명
        example: Legacy system user management API
      base_url:
        type: string
        description: 기본 URL
        example: https://legacy-api.example.com
      path:
        type: string
        description: API 경로
        example: /api/users
      health_url:
        type: string
        description: 헬스체크 URL
        example: https://legacy-api.example.com/health
      method:
        type: string
        description: HTTP 메서드
        example: GET
        enum:
          - GET
          - POST
          - PUT
          - DELETE
          - PATCH
      is_active:
        type: boolean
        description: 활성화 여부
        example: true
      timeout:
        type: integer
        description: 타임아웃 (밀리초)
        example: 30000
      retry_count:
        type: integer
        description: 재시도 횟수
        example: 3
      priority:
        type: integer
        description: 우선순위
        example: 1
  UpdateEndpointRequest:
    type: object
    properties:
      name:
        type: string
        description: 엔드포인트 이름
      description:
        type: string
        description: 엔드포인트 설명
      base_url:
        type: string
        description: 기본 URL
      health_url:
        type: string
        description: 헬스체크 URL
      is_active:
        type: boolean
        description: 활성화 여부
      timeout:
        type: integer
        description: 타임아웃 (밀리초)
      retry_count:
        type: integer
        description: 재시도 횟수
  EndpointResponse:
    type: object
    properties:
      id:
        type: string
        description: 엔드포인트 ID
        example: endpoint-20240101120000-abc123
      name:
        type: string
        description: 엔드포인트 이름
        example: Legacy User API
      description:
        type: string
        description: 엔드포인트 설명
      base_url:
        type: string
        description: 기본 URL
        example: https://legacy-api.example.com
      health_url:
        type: string
        description: 헬스체크 URL
      is_active:
        type: boolean
        description: 활성화 여부
        example: true
      timeout:
        type: integer
        description: 타임아웃 (밀리초)
        example: 30000
      retry_count:
        type: integer
        description: 재시도 횟수
        example: 3
      created_at:
        type: string
        format: date-time
        description: 생성 시간
      updated_at:
        type: string
        format: date-time
        description: 수정 시간
...
  ErrorResponse:
    type: object
    properties:
      error:
        type: string
        description: 에러 메시지
        example: endpoint not found
      details:
        type: string
        description: 상세 에러 내용
        example: no endpoint found with ID endpoint-123